// src/app/models/product.ts
export class product {
    "id": number;
    "name": string;
    "banner_img": string;
    "cover_img": string;
    "price": number;
    "sale": number;
    "description": string;
    "availability": number;
    "unit": string;
    "storage_instructions": string;
    "view": number;
    "revenue": number;
    "returned": number;
    "status": number;
    "seller_id": string;
    "category_id": number;
    "create_at": string;
    "update_at": string;
    "quantity": number
}